+++ R/utility_functions.R
@@ -83,6 +83,7 @@ dir_create_pipe <- function(path, showWarnings = TRUE, recursive = FALSE, mode =
 #' @param how_far The last target to include in the drake plan. Default is "all_targets"
 #' @param iea_data_path The path to IEA data. Default is `IEATools::sample_iea_data_path()`.
 #' @param fu_analysis_folder The path to the final-to-useful analysis folder. Default is `tempdir()`.
+#' @param reports_output_folder The path into which reports will be written. Default is `tempdir()`.
 #' @param exemplar_folder The path to a temporary folder to contain an exemplar table. Default is `tempdir()`.
 #' @param cache_path The path to the temporary drake cache used for testing. Default is `tempfile("drake_cache_for_testing")`.
 #' @param setup_exemplars Tells whether GHA allocation data should be adjusted to allow exemplars and
@@ -100,6 +101,8 @@ set_up_for_testing <- function(countries = c("GHA", "ZAF"),
                                how_far = "all_targets",
                                iea_data_path = IEATools::sample_iea_data_path(),
                                fu_analysis_folder = tempdir(),
+                               reports_source_folders = system.file("reports", package = "SEAPSUTWorkflow"),
+                               reports_output_folder = tempdir(),
                                exemplar_folder = tempdir(),
                                cache_path = tempfile("drake_cache_for_testing"),
                                setup_exemplars = FALSE) {
@@ -108,14 +111,16 @@ set_up_for_testing <- function(countries = c("GHA", "ZAF"),
   if (setup_exemplars & is.null(additional_exemplar_countries)) {
     additional_exemplar_countries = "World"
   }
-  set_up_temp_analysis(fu_analysis_folder, exemplar_folder, iea_data_path, setup_exemplars = setup_exemplars)
+  set_up_temp_analysis(fu_analysis_folder, exemplar_folder, reports_output_folder, iea_data_path, setup_exemplars = setup_exemplars)
   plan <- get_plan(countries = countries,
                    additional_exemplar_countries = additional_exemplar_countries,
                    max_year = max_year,
                    how_far = how_far,
                    iea_data_path = file.path(fu_analysis_folder, "IEAData.csv"),
                    exemplar_table_path = file.path(exemplar_folder, "Exemplar_Table.xlsx"),
-                   fu_analysis_folder = fu_analysis_folder)
+                   fu_analysis_folder = fu_analysis_folder,
+                   reports_source_folders = reports_source_folders,
+                   reports_dest_folder = reports_output_folder)
   temp_cache <- drake::new_cache(path = cache_path)
   list(plan = plan, cache_path = cache_path, temp_cache = temp_cache)
 }
@@ -132,6 +137,7 @@ set_up_for_testing <- function(countries = c("GHA", "ZAF"),
 #'
 #' @param fu_folder The folder in which the final-to-useful analysis structure will be created.
 #' @param exemplar_folder The folder in which a small exemplar table will be created.
+#' @param reports_output_folder The folder into which reports will be written.
 #' @param iea_data_path The path to an IEA data file.
 #' @param setup_exemplars Tells whether GHA allocation data should be adjusted to allow exemplars and
 #'                        if ZAF allocations will be duplicated and called "World".
@@ -140,7 +146,7 @@ set_up_for_testing <- function(countries = c("GHA", "ZAF"),
 #' @return `NULL`. This function should be called for its side effect of creating a temporary final-to-useful directory structure.
 #'
 #' @noRd
-set_up_temp_analysis <- function(fu_folder, exemplar_folder, iea_data_path, setup_exemplars = FALSE) {
+set_up_temp_analysis <- function(fu_folder, exemplar_folder, reports_output_folder, iea_data_path, setup_exemplars = FALSE) {
   # Set up IEA data
   iea_df <- IEATools::iea_df(iea_data_path)
   if (setup_exemplars) {
@@ -156,6 +162,10 @@ set_up_temp_analysis <- function(fu_folder, exemplar_folder, iea_data_path, setu
   # Write the iea_df to disk in the temporary directory (fu_folder)
   dir.create(fu_folder, showWarnings = FALSE)
   utils::write.csv(iea_df, file.path(fu_folder, "IEAData.csv"), row.names = FALSE)
+
+  # Create other temporary folders
+  dir.create(exemplar_folder, showWarnings = FALSE)
+  dir.create(reports_output_folder, showWarnings = FALSE)
 
   # Set up FU allocation tables. Need to split file that is stored in IEATools.
   GHAZAF_FU_allocation_tables <- IEATools::sample_fu_allocation_table_path() %>%